12/05 댓글을 제외한 모든 작업 완료

+++ src/main/resources/static/js/view.js
@@ -1,27 +1,44 @@
+var updateBtn = document.getElementById('updateBtn');
+var deleteBtn = document.getElementById('deleteBtn');
+var id = document.getElementById('num').value;
+var goHomeBtn = document.getElementById('goHome');
 
-function goUpdate(id){
-    location.replace(`/board/update?id=${id}`);
+if(updateBtn){
+
+    updateBtn.addEventListener('click',()=>{
+        location.replace(`/board/update?id=${id}`);
+    })
+}
+
+if(goHomeBtn){
+
+    goHomeBtn.addEventListener('click',()=>{
+        location.replace('/board/main');
+    });
 }
 
 
-function goDelete(id){
+if(deleteBtn){
 
-        fetch(`/api/board/${id}`,{
-            method:'DELETE'
-        })
-        .then(response =>{
-            if(!response.ok){
-                alert('게시물 삭제에 실패하였습니다.');
-                throw new Error(`HTTP error! Status: ${response.status}`);
-            }
-            else{
-                alert('삭제되었습니다.');
-                location.replace('/board/main');
-            }
-        })
-        .catch(error =>{
+        deleteBtn.addEventListener('click',()=>{
 
-            console.log('error',error);
+            fetch(`/api/board/${id}`,{
+                method:'DELETE'
+            })
+            .then(response =>{
+                if(!response.ok){
+                    alert('게시물 삭제에 실패하였습니다.');
+                    throw new Error(`HTTP error! Status: ${response.status}`);
+                }
+                else{
+                    alert('삭제되었습니다.');
+                    location.replace('/board/main');
+                }
+            })
+            .catch(error =>{
+
+                console.log('error',error);
+            })
         })
